refactor : 타입 정의 파일명 변경 오류 해결
