+++ src/entities/professor/model/professors.ts
